requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
bleach==3.3.1
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.3.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
flake8==7.0.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.9.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
packaging==24.0
phonenumbers==8.13.30
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
